Added separate controllers to separate version_one and version_two
