+++ app/helpers/version_one_helper.rb
@@ -1,0 +1,2 @@
+module VersionOneHelper
+end
