+++ app/helpers/version_two_helper.rb
@@ -1,0 +1,2 @@
+module VersionTwoHelper
+end
