+++ test/controllers/version_one_controller_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class VersionOneControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
